+++ bin/check-rabbitmq-node-health.rb
@@ -1,5 +1,6 @@
 #!/usr/bin/env ruby
 #  encoding: UTF-8
+
 #
 # RabbitMQ check node health plugin
 # ===
@@ -29,6 +30,7 @@ require 'sensu-plugin/check/cli'
 require 'json'
 require 'rest_client'
 require 'inifile'
+require 'socket'
 
 # main plugin class
 class CheckRabbitMQNodeHealth < Sensu::Plugin::Check::CLI
@@ -55,6 +57,21 @@ class CheckRabbitMQNodeHealth < Sensu::Plugin::Check::CLI
          short: '-P',
          long: '--port PORT',
          default: '15672'
+
+  option :node_prefix,
+         description: 'RabbitMQ node prefix (eg. rabbit in rabbit@hostname)',
+         long: '--prefix NODENAME',
+         default: 'rabbit'
+
+  option :node_use_fqdn,
+         description: 'Use FQDN for determining node hostname',
+         long: '--fqdn',
+         boolean: true,
+         default: false
+
+  option :node_host,
+         description: 'RabbitMQ node name override (eg. hostname in rabbit@hostname)',
+         long: '--nodehost HOSTNAME'
 
   option :ssl,
          description: 'Enable SSL for connection to the API',
@@ -136,12 +153,26 @@ class CheckRabbitMQNodeHealth < Sensu::Plugin::Check::CLI
   end
 
   def node_healthy?
-    host       = config[:host]
-    port       = config[:port]
-    username   = config[:username]
-    password   = config[:password]
-    ssl        = config[:ssl]
-    verify_ssl = config[:verify_ssl_off]
+    host        = config[:host]
+    port        = config[:port]
+    username    = config[:username]
+    password    = config[:password]
+    ssl         = config[:ssl]
+    verify_ssl  = config[:verify_ssl_off]
+    node_prefix = config[:node_prefix]
+
+    # Determine node hostname to query, as this may not be the same as connection hostname
+    node_host = if config[:node_host]
+                  # Statically set by config
+                  config[:node_host]
+                elsif config[:node_use_fqdn]
+                  # Resolved from system running check, full FQDN, the same as RABBITMQ_USE_LONGNAME
+                  Socket.gethostname
+                else
+                  # Default of short hostname, resolved from system
+                  Socket.gethostname.partition('.').first
+                end
+
     if config[:ini]
       ini = IniFile.load(config[:ini])
       section = ini['auth']
@@ -154,14 +185,15 @@ class CheckRabbitMQNodeHealth < Sensu::Plugin::Check::CLI
 
     begin
       url_prefix = ssl ? 'https' : 'http'
+      node = "#{node_prefix}@#{node_host}"
       resource = RestClient::Resource.new(
-        "#{url_prefix}://#{host}:#{port}/api/nodes",
+        "#{url_prefix}://#{host}:#{port}/api/nodes/#{node}",
         user: username,
         password: password,
         verify_ssl: !verify_ssl
       )
       # Parse our json data
-      nodeinfo = JSON.parse(resource.get)[0]
+      nodeinfo = JSON.parse(resource.get)
 
       # Determine % memory consumed
       pmem = format('%.2f', nodeinfo['mem_used'].fdiv(nodeinfo['mem_limit']) * 100)